Add CreateMonster and CreatePlayer components
